refactor(Tag): split out standalone methods to trivial base
